+++ fennel/CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## [0.18.1] - 2023-09-19
+## [0.18.4] - 2023-09-19
 - Added select operator
+
+## [0.18.2] - 2023-09-13
+- Add derived extractors for aliasing and lookups
+
+## [0.18.1] - 2023-09-08
+- Add support to specify output bucket and prefix for extract historical, and support to map output columns to different features.
 
 ## [0.18.0] - 2023-08-30
 - Added support for Debezium data in Avro format via Kafka connector